Move create_root_admin out of main.py
