+++ app/admin/__init__.py
@@ -1,2 +1,2 @@
 """init package"""
-from .admin import create_admin
+from .admin import create_admin, create_root_admin
